refactor: updated folder name
